authentication added

+++ src/main/java/com/edstem/taxibookingsystem/model/User.java
@@ -28,6 +28,6 @@ public class User {
     private String name;
     private String email;
     private String password;
-    private Long accountBalance;
+    private Double accountBalance;
 
 }
